requirements.txt
Django==1.11.7
django-appconf==1.0.2
django-imagekit==4.0.1
olefile==0.44
pilkit==2.0
Pillow==4.3.0
pytz==2017.3
six==1.11.0
